refactor: Add apiCalls.js file for separation of concerns
